EIDOLON-V Phase 3: Zero-Allocation Networking & DOD Optimization
## Major Changes

### 1. SchemaBinaryPacker - Pre-encoded Component IDs
- Pre-encode component IDs (STATS, SKILL, PIGMENT, etc.) as static Uint8Array
- Eliminates TextEncoder allocation in hot loop (60Hz x N players)
- Zero-allocation networking path

### 2. ServerValidator - Composite Key Architecture
- Replace string concatenation `_` with nested Map
- Map<string, Map<string, RateTracker>> eliminates string garbage
- O(1) lookup without heap allocation

### 3. PhysicsSystem - Embedded Anti-cheat
- Add MAX_SPEED_BASE and SPEED_VALIDATION_TOLERANCE constants
- Speed clamping integrated directly in integrateEntity()
- Server-authoritative velocity limits at physics level

### 4. GameRoom - syncDODToSchema Elimination
- **REMOVED** syncDODToSchema() method entirely
- Position/Velocity now exclusively via binary snapshots
- checkPlayerDeaths() handles death state sync only
- 4-step update loop reduced to: input → physics → logic → binary broadcast

### 5. Build System
- Deleted BinaryPacker.ts (deprecated)
- Created compat.ts as DOD compatibility layer
- All imports migrated to generated/ or compat/
- Build passes: tsc --noEmit ✓

## Performance Impact
- TextEncoder allocations: ELIMINATED
- String concatenation in hot path: ELIMINATED
- Colyseus state writes per player: REDUCED from 8 to 1 (only death state)
- syncDODToSchema overhead: ELIMINATED

## Files Modified
- packages/engine/src/networking/SchemaBinaryPacker.ts
- packages/engine/src/systems/PhysicsSystem.ts
- apps/server/src/security/ServerValidator.ts
- apps/server/src/rooms/GameRoom.ts
- packages/engine/src/compat.ts (new)
- packages/engine/src/dod/* → removed (migrated to compat)
- packages/engine/src/networking/BinaryPacker.ts → deleted

+++ packages/engine/scripts/schema.config.js
@@ -74,7 +74,7 @@ const SCHEMA = {
         fields: {
             targetX: { type: 'f32', default: 0 },
             targetY: { type: 'f32', default: 0 },
-            actions: { type: 'f32', default: 0 }, // Bitmask stored as float for TypedArray compat
+            actions: { type: 'u32', default: 0 }, // EIDOLON-V FIX: Uint32 for accurate bitmask
             _pad: { type: 'f32', default: 0 },
         }
     },
